requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
asttokens==3.0.0
attrs==25.3.0
certifi==2025.10.5
charset-normalizer==3.4.3
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
Django==5.2.7
django-environ==0.12.0
django-markdownx==4.0.9
executing==2.2.1
faiss-cpu==1.12.0
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.78
langchain-openai==0.3.34
langchain-text-splitters==0.3.11
langsmith==0.4.32
Markdown==3.9
markdown-katex==202406.1035
marshmallow==3.26.1
matplotlib-inline==0.1.7
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.3
openai==2.1.0
orjson==3.11.3
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.4.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
pylatexenc==2.10
PyMuPDF==1.26.4
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
stack-data==0.6.3
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
yarl==1.20.1
zstandard==0.25.0
